[AIRFLOW-1314] fixes after rebase
