+++ airflow/contrib/kubernetes/pod.py
@@ -1,20 +1,29 @@
-# -*- coding: utf-8 -*-
+
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
 #
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
 
 
 class Resources:
-    def __init__(self, request_memory=None, request_cpu=None, limit_memory=None, limit_cpu=None):
+    def __init__(
+            self,
+            request_memory=None,
+            request_cpu=None,
+            limit_memory=None,
+            limit_cpu=None):
         self.request_memory = request_memory
         self.request_cpu = request_cpu
         self.limit_memory = limit_memory
@@ -33,7 +42,6 @@ class Resources:
 class Pod:
     """
     Represents a kubernetes pod and manages execution of a single pod.
-
     :param image: The docker image
     :type image: str
     :param env: A dict containing the environment variables
@@ -46,18 +54,18 @@ class Pod:
                    successful execution of the pod
     :type result: any
     """
-    pod_timeout = 3600
-
     def __init__(
             self,
             image,
             envs,
             cmds,
-            secrets,
+            args=None,
+            secrets=None,
             labels=None,
             node_selectors=None,
             name=None,
-            volumes = [],
+            volumes=None,
+            volume_mounts=None,
             namespace='default',
             result=None,
             image_pull_policy="IfNotPresent",
@@ -67,14 +75,16 @@ class Pod:
             resources=None
     ):
         self.image = image
-        self.envs = envs if envs else {}
+        self.envs = envs or {}
         self.cmds = cmds
-        self.secrets = secrets
+        self.args = args or []
+        self.secrets = secrets or []
         self.result = result
-        self.labels = labels if labels else []
+        self.labels = labels or {}
         self.name = name
-        self.volumes = volumes
-        self.node_selectors = node_selectors if node_selectors else []
+        self.volumes = volumes or []
+        self.volume_mounts = volume_mounts or []
+        self.node_selectors = node_selectors or []
         self.namespace = namespace
         self.image_pull_policy = image_pull_policy
         self.image_pull_secrets = image_pull_secrets
