+++ airflow/contrib/kubernetes/pod_launcher.py
@@ -1,28 +1,31 @@
-# -*- coding: utf-8 -*-
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
 #
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
 
 import json
-
-from airflow.contrib.kubernetes.pod import Pod
-from airflow.contrib.kubernetes.kubernetes_request_factory.pod_request_factory import (
-    SimplePodRequestFactory)
+import time
 from airflow.utils.log.logging_mixin import LoggingMixin
 from airflow.utils.state import State
+from datetime import datetime as dt
+from airflow.contrib.kubernetes.kubernetes_request_factory import \
+    pod_request_factory as pod_fac
 from kubernetes import watch
-from kubernetes.client import V1Pod
 from kubernetes.client.rest import ApiException
-
+from airflow import AirflowException
+from requests.exceptions import HTTPError
 from .kube_client import get_kube_client
 
 
@@ -35,9 +38,10 @@ class PodStatus(object):
 
 class PodLauncher(LoggingMixin):
     def __init__(self, kube_client=None):
-        self.kube_req_factory = SimplePodRequestFactory()
+        super(PodLauncher, self).__init__()
         self._client = kube_client or get_kube_client()
         self._watch = watch.Watch()
+        self.kube_req_factory = pod_fac.SimplePodRequestFactory()
 
     def run_pod_async(self, pod):
         req = self.kube_req_factory.create(pod)
@@ -50,35 +54,61 @@ class PodLauncher(LoggingMixin):
             raise
         return resp
 
-    def run_pod(self, pod):
+    def run_pod(self, pod, startup_timeout=120):
         # type: (Pod) -> State
         """
         Launches the pod synchronously and waits for completion.
+        Args:
+            pod (Pod):
+            startup_timeout (int): Timeout for startup of the pod (if pod is pending for
+             too long, considers task a failure
         """
         resp = self.run_pod_async(pod)
+        curr_time = dt.now()
+        if resp.status.start_time is None:
+            while self.pod_not_started(pod):
+                delta = dt.now() - curr_time
+                if delta.seconds >= startup_timeout:
+                    raise AirflowException("Pod took too long to start")
+                time.sleep(1)
+            self.log.debug('Pod not yet started')
+
         final_status = self._monitor_pod(pod)
         return final_status
 
     def _monitor_pod(self, pod):
         # type: (Pod) -> State
-        for event in self._watch.stream(self.read_pod(pod), pod.namespace):
-            status = self._task_status(event)
-            if status == State.SUCCESS or status == State.FAILED:
-                return status
+
+        while self.pod_is_running(pod):
+            self.log.info("Pod {} has state {}".format(pod.name, State.RUNNING))
+            time.sleep(2)
+        return self._task_status(self.read_pod(pod))
 
     def _task_status(self, event):
         # type: (V1Pod) -> State
-        task = event['object']
         self.log.info(
-            "Event: {} had an event of type {}".format(task.metadata.name,
-                                                       event['type']))
-        status = self.process_status(task.metadata.name, task.status.phase)
+            "Event: {} had an event of type {}".format(event.metadata.name,
+                                                       event.status.phase))
+        status = self.process_status(event.metadata.name, event.status.phase)
         return status
 
+    def pod_not_started(self, pod):
+        state = self._task_status(self.read_pod(pod))
+        return state == State.QUEUED
+
+    def pod_is_running(self, pod):
+        state = self._task_status(self.read_pod(pod))
+        return state != State.SUCCESS and state != State.FAILED
+
     def read_pod(self, pod):
-        return self._client.read_namespaced_pod(pod.name, pod.namespace)
+        try:
+            return self._client.read_namespaced_pod(pod.name, pod.namespace)
+        except HTTPError as e:
+            raise AirflowException("There was an error reading the kubernetes API: {}"
+                                   .format(e))
 
     def process_status(self, job_id, status):
+        status = status.lower()
         if status == PodStatus.PENDING:
             return State.QUEUED
         elif status == PodStatus.FAILED:
