+++ scripts/ci/kubernetes/docker/Dockerfile
@@ -33,6 +33,9 @@ RUN apt-get update -y && apt-get install -y \
 RUN pip install -U setuptools && \
     pip install -U pip
 
+COPY requirements.txt /tmp/requirements.txt
+RUN pip install -r /tmp/requirements.txt
+
 RUN pip install kubernetes && \
     pip install cryptography && \
     pip install psycopg2==2.7.3.1  # I had issues with older versions of psycopg2, just a warning
