+++ scripts/ci/kubernetes/docker/requirements.txt
@@ -1,0 +1,35 @@
+alembic
+bleach
+configparser
+croniter
+dill
+flask
+flask-admin
+flask-caching
+flask-login
+flask-swagger
+flask-wtf
+funcsigs
+future
+gitpython
+gunicorn
+iso8601
+jinja2
+lxml
+markdown
+pandas
+pendulum
+psutil
+pygments
+python-daemon
+python-dateutil
+python-nvd3
+requests
+setproctitle
+sqlalchemy
+sqlalchemy-utc
+tabulate
+thrift
+tzlocal
+werkzeug
+zope.deprecation
